Added support for importing connections.csv, and bug fixes
You can now add connections from just dragging the connection.csv file into the workspace and setting .env var for scraper_mode to connections. It will look at the linkedin url, scrape their info, if they go/went to UNT. If not, it disregards as the connections might be professors or other people.